refactor: use oclif exactlyOne, hoist methods to deployCommand
